refactor: make page file names consistent
